chore(hoggar): adjusted folder structure
