+++ hoggar/src/index.ts
@@ -1,5 +1,5 @@
 import app from "./app";
-import { hostname, port } from "./config";
+import { hostname, port } from "./app/config";
 
 app.listen(port, hostname, () => {
   console.log(`⚡️[server]: Server is running at http://${hostname}:${port}`);
